feat: rename aws tracing module
