+++ observability-tracing-aws-xray/.dockerignore
