+++ observability-tracing-aws-xray/Dockerfile
