+++ observability-tracing-aws-xray/helm/templates/adapter/serviceaccount.yaml
@@ -5,10 +5,10 @@
 apiVersion: v1
 kind: ServiceAccount
 metadata:
-  name: tracing-adapter-cloudwatch
+  name: tracing-adapter-aws-xray
   namespace: {{ .Release.Namespace }}
   labels:
-    app: tracing-adapter-cloudwatch
+    app: tracing-adapter-aws-xray
   {{- with .Values.adapter.serviceAccount.annotations }}
   annotations:
     {{- toYaml . | nindent 4 }}
