+++ observability-tracing-aws-xray/helm/templates/aws-credentials/secret.yaml
@@ -2,7 +2,7 @@
 Static AWS credentials Secret for the k3d / kind path. On EC2 / EKS with an
 instance-profile, IRSA, or Pod Identity, leave `awsCredentials.create=false`.
 */ -}}
-{{- include "tracing-cloudwatch.validate" . }}
+{{- include "tracing-aws-xray.validate" . }}
 {{- if .Values.awsCredentials.create }}
 {{- if not .Values.awsCredentials.accessKeyId }}
 {{- fail "awsCredentials.accessKeyId is required when awsCredentials.create=true" }}
@@ -16,10 +16,10 @@ metadata:
   name: {{ .Values.awsCredentials.name }}
   namespace: {{ .Release.Namespace }}
   labels:
-    app.kubernetes.io/name: observability-tracing-cloudwatch
+    app.kubernetes.io/name: observability-tracing-aws-xray
 type: Opaque
 stringData:
   AWS_ACCESS_KEY_ID: {{ .Values.awsCredentials.accessKeyId | quote }}
   AWS_SECRET_ACCESS_KEY: {{ .Values.awsCredentials.secretAccessKey | quote }}
-  AWS_REGION: {{ include "tracing-cloudwatch.region" . | quote }}
+  AWS_REGION: {{ include "tracing-aws-xray.region" . | quote }}
 {{- end }}
