+++ observability-tracing-aws-xray/internal/api/cfg-models.yaml
