+++ observability-tracing-aws-xray/internal/api/cfg-server.yaml
