+++ observability-tracing-aws-xray/internal/handlers_test.go
@@ -10,8 +10,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/openchoreo/community-modules/observability-tracing-cloudwatch/internal/api/gen"
-	"github.com/openchoreo/community-modules/observability-tracing-cloudwatch/internal/xray"
+	"github.com/openchoreo/community-modules/observability-tracing-aws-xray/internal/api/gen"
+	"github.com/openchoreo/community-modules/observability-tracing-aws-xray/internal/xray"
 )
 
 func testLogger() *slog.Logger {
