+++ observability-tracing-aws-xray/internal/xray/client_test.go
