--- observability-tracing-cloudwatch/helm/templates/collector-env/configmap.yaml
@@ -1,13 +1,0 @@
-# Copyright 2026 The OpenChoreo Authors
-# SPDX-License-Identifier: Apache-2.0
-
-{{- include "tracing-cloudwatch.validate" . }}
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: tracing-cloudwatch-collector-env
-  namespace: {{ .Release.Namespace }}
-  labels:
-    app.kubernetes.io/name: observability-tracing-cloudwatch
-data:
-  AWS_REGION: {{ include "tracing-cloudwatch.region" . | quote }}
